+++ src/main/java/com/ms/service/model/ServiceModel.java
@@ -9,6 +9,7 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -22,7 +23,7 @@ public class ServiceModel implements Serializable {
     private String serviceEmail; // email
     private boolean serviceStatus; // status
     private String serviceDescription; // descrição
-    private double servicePrice; // preço
+    private BigDecimal servicePrice; // preço
     private int serviceRuntime; // tempo de execução
     private int serviceTerm; // prazo
     private String created;
